Add invalid roomId edge case
* The /room/join method could end up in an error if the user enters an
  invalid roomId. The blueprint to handle this path was setup.

* The HTML page to show the user when this path occurs, was also added.
