+++ src/components/Freepik404.astro
@@ -1,0 +1,37 @@
+---
+---
+
+<div class="freepik-404-container">
+  <img src="/talki/404.svg" />
+  <a
+    class="freepik-404-img font-mono"
+    href="https://www.freepik.com/free-vector/404-error-with-person-looking-concept-illustration_20824298.htm#query=found&position=18&from_view=keyword&track=sph"
+    >Image by storyset on Freepik</a
+  >
+</div>
+<style>
+  .freepik-404-container {
+    display: flex;
+    position: relative;
+  }
+  .freepik-404-img {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    position: absolute;
+    height: 100%;
+    width: 100%;
+    opacity: 0;
+    background-color: rgba(0, 0, 0, 0);
+    color: white;
+    transition-property: background-color opacity;
+    transition-duration: 0.3s;
+    box-shadow: 0px 0px 2rem black;
+    padding: 1rem;
+    /* border-radius: 1.5rem; */
+  }
+  .freepik-404-img:hover {
+    opacity: 1;
+    background-color: rgba(0, 0, 0, 0.6);
+  }
+</style>
