+++ src/pages/room/404.astro
@@ -1,0 +1,20 @@
+---
+import Freepik404 from "components/Freepik404.astro";
+import BasicHTMLLayout from "layouts/BasicHTMLLayout.astro";
+import GoToHome from "scripts/components/GoToHome.astro";
+---
+
+<BasicHTMLLayout title="Invalid room">
+  <div
+    class="h-screen flex flex-col justify-center items-center mx-8 text-center"
+  >
+    <div class="w-4/5 max-w-lg mb-8">
+      <Freepik404 />
+    </div>
+    <h1 class="text-4xl font-bold mb-4">Invalid room ID</h1>
+    <p class="font-mono mb-4 text-lg">
+      Looks like that room doesn't exist. Try again with a valid room ID.
+    </p>
+    <GoToHome />
+  </div>
+</BasicHTMLLayout>
