+++ src/scripts/components/GoToHome.astro
@@ -1,0 +1,6 @@
+<a
+  href="/talki/"
+  class="bg-talki-black hover:bg-gray-300 focus-visible:bg-gray-300 hover:outline hover:outline-2 hover:outline-gray-500 px-4 py-2 group text-white font-mono font-bold hover:text-black"
+>
+  <span aria-hidden="true">&#8592</span> Go to Home Page
+</a>
