+++ src/scripts/pages/room/startSession/index.ts
@@ -1,13 +1,10 @@
-import roomCreateEndpoint, {
-  RoomCreateEndpoint,
-} from "utils/endpoints/room/create";
-import roomJoinEndpoint, { RoomJoinEndpoint } from "utils/endpoints/room/join";
-import backendFetch from "utils/functions/backendFetch";
+import { redirectToInternalUrl } from "utils/functions/redirects";
 import type { RoomId } from "utils/types/Room";
 import setRoomIdOnPage from "../pageManip/roomId";
 import addStreamContainer from "../pageManip/streamContainer";
 import { getRoomIdFromUrl, setRoomIdInUrl } from "../url";
 import connectRoomWs from "./connectRoomWs";
+import { createRoom, tryJoinRoom } from "./room";
 
 async function startSession() {
   const roomId = getRoomIdFromUrl();
@@ -18,34 +15,27 @@ async function startSession() {
   addStreamContainer(localStream);
 
   if (roomId) {
-    setRoomIdOnPage(roomId);
     joinExistingCall(roomId);
     return;
   }
   await startNewCall();
 }
 
-async function joinExistingCall(roomId: RoomId) {
-  const { url, headers, method } = roomJoinEndpoint;
-  const body: RoomJoinEndpoint["body"] = { roomId };
-  const res = await backendFetch(url, {
-    method,
-    headers,
-    body: JSON.stringify(body),
-  });
-  const { wsUrl, memberId, nonce, expireAt } =
-    (await res.json()) as RoomJoinEndpoint["response"];
-  await connectRoomWs({ wsUrl, roomId, memberId, nonce, expireAt });
-}
-
 async function startNewCall() {
-  const { url, method } = roomCreateEndpoint;
-  const res = await backendFetch(url, { method });
-  const { wsUrl, roomId, memberId, nonce, expireAt } =
-    (await res.json()) as RoomCreateEndpoint["response"];
+  const { wsUrl, roomId, memberId, nonce, expireAt } = await createRoom();
   setRoomIdInUrl(roomId);
   setRoomIdOnPage(roomId);
   await connectRoomWs({ wsUrl, roomId, memberId, nonce, expireAt });
+}
+
+async function joinExistingCall(roomId: RoomId) {
+  const joinAttempt = await tryJoinRoom(roomId);
+  if (!joinAttempt.success) {
+    redirectToInternalUrl("/room/404");
+    return;
+  }
+  setRoomIdOnPage(roomId);
+  await connectRoomWs(joinAttempt.data);
 }
 
 export default startSession;
